Fix broken blocking for vobadirekt.de
Using the block list makes the website inaccessible. The filters leave behind
an empty modal and the drop shadow that covers the page.

According to the uBlock Origin wiki [1] the :xpath selector only works
in Nano Blocker and uBlock Origin, but I couldn't find a better way to
write this query. The only alternative I see would be to always block
.darken-layer, regardless of which modal is shown.

[1] https://github.com/DandelionSprout/adfilt/blob/master/Wiki/SyntaxMeaningsThatAreActuallyHumanReadable.md#nano-adblocker-and-ublock-origin-only